requirements.txt
requests
python-dotenv
RocketChatAPIBot